# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Qubirex API",
    "version": "4.3.1-phase0.1",
    "description": "Generated from the Express routers by scripts/openapi.js. Do not edit by hand."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "cookieSession": []
    },
    {
      "bearer": []
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "qbx_session",
        "description": "httpOnly SameSite=Strict session cookie. State-changing requests also need X-CSRF-Token."
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Same session token; accepted until the frontend moves to cookies (docs/decisions.md D-002)."
      }
    },
    "parameters": {
      "CsrfToken": {
        "name": "X-CSRF-Token",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Required when authenticating with the session cookie."
      }
    },
    "responses": {
      "Error": {
        "description": "Error. New routes (/api/employer, /api/verify, /api/practical) answer {error: {code, message}}; legacy routes answer {error: string}.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "oneOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    }
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/learner/learning-curve": {
      "get": {
        "tags": [
          "learner"
        ],
        "operationId": "get_learner_learning_curve",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/learner/passport": {
      "get": {
        "tags": [
          "learner"
        ],
        "operationId": "get_learner_passport",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/learner/passport/renew": {
      "post": {
        "tags": [
          "learner"
        ],
        "operationId": "post_learner_passport_renew",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ]
      }
    },
    "/api/learner/passport/share": {
      "post": {
        "tags": [
          "learner"
        ],
        "operationId": "post_learner_passport_share",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ]
      }
    },
...
    "/api/learner/reviews/due": {
      "get": {
        "tags": [
          "learner"
        ],
        "operationId": "get_learner_reviews_due",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/learner/reviews/instances/{id}/answer": {
      "post": {
        "tags": [
          "learner"
        ],
        "operationId": "post_learner_reviews_instances_id_answer",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ]
      }
    },
    "/api/learner/reviews/{nodeId}/start": {
      "post": {
        "tags": [
          "learner"
        ],
        "operationId": "post_learner_reviews_nodeId_start",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "nodeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ]
      }
    },
...
  }
}